+++ park-here-server/src/regions/region.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 // must be renamed to location
 #[derive(Serialize)]
 pub struct Region {
-    pub latitude: f32,
-    pub longitude: f32,
+    pub latitude: f64,
+    pub longitude: f64,
     pub id: String,
 }
